Moved Window header and implementation to rest of windows.
