+++ CMakeLists.txt
@@ -87,7 +87,6 @@ set(srcFileList
 	${srcGraphicDir}/DynamicObject.cpp
 	${srcGraphicDir}/Aim.cpp
 	${srcGraphicDir}/Camera.cpp
-	${srcGraphicDir}/Window.cpp
 	${srcGraphicDir}/Particle.cpp
 	${srcGraphicDir}/ParticleSystem.cpp
 	${srcGraphicDir}/Explosion.cpp
@@ -96,6 +95,7 @@ set(srcFileList
 	${srcTextDir}/TextObject.cpp 
 	${srcSoundDir}/Sound.cpp
 	${srcSoundDir}/SoundManager.cpp
+	${srcWindowsDir}/Window.cpp
 	${srcWindowsDir}/MainMenuWindow.cpp
 	${srcWindowsDir}/CharStatsWindow.cpp
 )
