+++ src/windows/MainMenuWindow.h
@@ -1,7 +1,7 @@
 #ifndef MAINMENUWINDOW_H_
 #define MAINMENUWINDOW_H_
 
-#include "../system/graphic/Window.h"
+#include "Window.h"
 #include "../system/Configuration.h"
 #include "../game/GameState.h"
 #include "../system/graphic/text/TextManager.h"
@@ -18,7 +18,7 @@ public:
 
 	MainMenuWindow(Configuration* config, GameState* gameState,
 			TextManager* text);
-	~MainMenuWindow();
+	virtual ~MainMenuWindow();
 };
 }
